not use the off local variable
It has to do with the "off" local variable and the fact that it's reused for different attributes. The optimization problem boils down to the compiler getting confused in this code pattern about the different values that are stored into "off" and then used as part of handling that particular attribute. That confusion prevents folding what should be constant expressions in quite a few places.

Solution: not use the off local variable and instead just consume the computed value that we store into that variable directly. That may mean passing around the TR::IlValue *value that you passed to Store("off", value) and simply passing value to subsequent operations rather than doing Load("off").